extract method fitToDst
