+++ src/js/utils/index.js
@@ -182,6 +182,33 @@ const truncateMiddle = (string, maxLength = 30, separator = '…') => {
 
 const dashed = str => str.replace(/\s+/g, '-')
 
+// via https://stackoverflow.com/questions/6565703/math-algorithm-fit-image-to-screen-retain-aspect-ratio
+//
+// Image data: (wi, hi) and define ri = wi / hi
+// Screen resolution: (ws, hs) and define rs = ws / hs
+//
+// rs > ri ? (wi * hs/hi, hs) : (ws, hi * ws/wi)
+//
+// top = (hs - hnew)/2
+// left = (ws - wnew)/2
+
+const fitToDst = (dst, src) => {
+  let wi = src.width
+  let hi = src.height
+  let ri = wi / hi
+
+  let ws = dst.width
+  let hs = dst.height
+  let rs = ws / hs
+
+  let [wnew, hnew] = rs > ri ? [wi * hs / hi, hs] : [ws, hi * ws / wi]
+
+  let x = (ws - wnew) / 2
+  let y = (hs - hnew) / 2
+
+  return [x, y, wnew, hnew]
+}
+
 module.exports = {
   msToTime,
   uidGen,
@@ -201,5 +228,6 @@ module.exports = {
   zeroFill,
   uniq,
   truncateMiddle,
-  dashed
+  dashed,
+  fitToDst
 }
